api/pyproject.toml
[tool.poetry]
name = "filemanager-fastapi"
version = "0.0.1"
description = "Fastapi with file management"
authors = ["JEX <guja1995@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.60.1"
uvicorn = "^0.11.8"
gunicorn = "^20.0.4"
python-multipart = "^0.0.5"
python-dotenv = "^0.14.0"
apache-libcloud = "^3.1.0"
Pillow-SIMD = "^7.0.0"
aiofiles = "^0.5.0"
cryptography = "^3.0"
ffmpeg-python = "^0.2.0"
qrcode = "^6.1"

[tool.poetry.dev-dependencies]
pylint = "^2.5.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
